test: use activity params test
