+++ packages/graphql-client/documents/delegate/drawer/voting-power/delegators.graphql
@@ -1,8 +1,8 @@
 query GetDelegators(
   $address: String!
   $skip: Int
-  $limit: Int = 10
   $orderBy: queryInput_delegators_orderBy = amount
+  $limit: Int = 10
   $orderDirection: OrderDirection = desc
 ) {
   delegators(
